backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
openai==1.52.2
httpx==0.27.2
httpcore==1.0.5
pillow==12.0.0
reportlab==4.4.4
pydantic==2.9.2
pydantic_core==2.23.4
starlette==0.40.0
python-multipart==0.0.9
